+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Implemented the pf-table component.
+
 ### Fixed
 - Fixed display of column picker in the pf-toolbar when the sort and/or list controls are disabled.